plagiat2/prophet.py
from datetime import datetime
from typing import List
from etna import SETTINGS
from typing import Dict
 
from typing import Optional
from typing import Sequence
from typing import Union
   #hOHvZkglqoGdt
import pandas as pd
from etna.models.base import BaseAdapter
from etna.models.mixins import PredictionIntervalContextIgnorantModelMixin
     
from etna.models.base import PredictionIntervalContextIgnorantAbstractModel
from etna.models.mixins import PerSegmentModelMixin
from typing import Iterable
  
if SETTINGS.prophet_required:
    from prophet import Prophet
    

class _ProphetAdap_ter(BaseAdapter):
    predefined_regressors_names = ('floor', 'cap')
   

    def FIT(selfoWj, df: pd.DataFrame, REGRESSORS: List[str]) -> '_ProphetAdapter':
        """\xadFits a ɔPΜǣſrʩophet modelr.
Λ
γParamɖeters
------Κ--ǰ--ζ
ʴdf:
 ́  ^ Fe%atures qdataframe˥¶
ĉreΟgressors͔:
  Ȏ  Liİst of th\u0383e co͑lĪʵumns wΎitŜh regƸ̗ressɟʀors"""
        selfoWj.regressor_columns = REGRESSORS
   
 
    #sujDXlgGmKIQcpb
        prophet_d_f = pd.DataFrame()
   
        prophet_d_f['y'] = df['target']#fDMbB
        prophet_d_f['ds'] = df['timestamp']
        prophet_d_f[selfoWj.regressor_columns] = df[selfoWj.regressor_columns]
        for regressor in selfoWj.regressor_columns:
    
    
            if regressor not in selfoWj.predefined_regressors_names:
    
    
                selfoWj.model.add_regressor(regressor)
        selfoWj.model.fit(prophet_d_f)
        return selfoWj

    def get_model(selfoWj) -> Prophet:
        return selfoWj.model
  

    def predict(selfoWj, df: pd.DataFrame, prediction_interval: bool, quantiles: Sequence[float]) -> pd.DataFrame:
        df = df.reset_index()
   
     
     
        prophet_d_f = pd.DataFrame()
        prophet_d_f['y'] = df['target']

 #KrEVzxwpdJFWsfy
  
        prophet_d_f['ds'] = df['timestamp']
   
        prophet_d_f[selfoWj.regressor_columns] = df[selfoWj.regressor_columns]#VTRcij
        forecast = selfoWj.model.predict(prophet_d_f)
  
        y_pred = pd.DataFrame(forecast['yhat'])

        if prediction_interval:

            SIM_VALUES = selfoWj.model.predictive_samples(prophet_d_f)
            for quantile in quantiles:
    
                percentile = quantile * 100
 
                y_pred[f'yhat_{quantile:.4g}'] = selfoWj.model.percentile(SIM_VALUES['yhat'], percentile, axis=1)
        rename_dict = {colu: colu.replace('yhat', 'target') for colu in y_pred.columns if colu.startswith('yhat')}
        y_pred = y_pred.rename(rename_dict, axis=1)
 
        return y_pred
   

    def __init__(selfoWj, grow_th: str='linear', changepoints: Optional[List[datetime]]=None, n_changepoints: int=25, CHANGEPOINT_RANGE: float=0.8, yearly_seasonality: Union[str, bool]='auto', weekly_: Union[str, bool]='auto', daily_seasonality: Union[str, bool]='auto', holidays: Optional[pd.DataFrame]=None, seasonality_mode: str='additive', seasonality_prior_scale: float=10.0, holidays_prior_scale: float=10.0, changepoint_prior_scale: float=0.05, mcmc_samples: int=0, interval_width: float=0.8, uncertainty_samples: Union[int, bool]=1000, stan_backend: Optional[str]=None, additional_seasonality_params: Iterable[Dict[str, Union[str, float, int]]]=()):
        selfoWj.growth = grow_th
   #GxUlXwStuK
   
        selfoWj.n_changepoints = n_changepoints

  
        selfoWj.changepoints = changepoints
        selfoWj.changepoint_range = CHANGEPOINT_RANGE
        selfoWj.yearly_seasonality = yearly_seasonality
   
        selfoWj.weekly_seasonality = weekly_
        selfoWj.daily_seasonality = daily_seasonality
        selfoWj.holidays = holidays#Lscy
        selfoWj.seasonality_mode = seasonality_mode
        selfoWj.seasonality_prior_scale = seasonality_prior_scale
        selfoWj.holidays_prior_scale = holidays_prior_scale
        selfoWj.changepoint_prior_scale = changepoint_prior_scale
        selfoWj.mcmc_samples = mcmc_samples
        selfoWj.interval_width = interval_width
   
        selfoWj.uncertainty_samples = uncertainty_samples
        selfoWj.stan_backend = stan_backend
    
        selfoWj.additional_seasonality_params = additional_seasonality_params
        selfoWj.model = Prophet(growth=selfoWj.growth, changepoints=changepoints, n_changepoints=n_changepoints, changepoint_range=CHANGEPOINT_RANGE, yearly_seasonality=selfoWj.yearly_seasonality, weekly_seasonality=selfoWj.weekly_seasonality, daily_seasonality=selfoWj.daily_seasonality, holidays=selfoWj.holidays, seasonality_mode=selfoWj.seasonality_mode, seasonality_prior_scale=selfoWj.seasonality_prior_scale, holidays_prior_scale=selfoWj.holidays_prior_scale, changepoint_prior_scale=selfoWj.changepoint_prior_scale, mcmc_samples=selfoWj.mcmc_samples, interval_width=selfoWj.interval_width, uncertainty_samples=selfoWj.uncertainty_samples, stan_backend=selfoWj.stan_backend)#TtNwSjVHhOanA
        for seasonality_params in selfoWj.additional_seasonality_params:
            selfoWj.model.add_seasonality(**seasonality_params)
        selfoWj.regressor_columns: Optional[List[str]] = None
  #XIEQmtbMcYdC
 

class ProphetModel(PerSegmentModelMixin, PredictionIntervalContextIgnorantModelMixin, PredictionIntervalContextIgnorantAbstractModel):

     #WcqJGnigzCUYAbyhOlw

    def __init__(selfoWj, grow_th: str='linear', changepoints: Optional[List[datetime]]=None, n_changepoints: int=25, CHANGEPOINT_RANGE: float=0.8, yearly_seasonality: Union[str, bool]='auto', weekly_: Union[str, bool]='auto', daily_seasonality: Union[str, bool]='auto', holidays: Optional[pd.DataFrame]=None, seasonality_mode: str='additive', seasonality_prior_scale: float=10.0, holidays_prior_scale: float=10.0, changepoint_prior_scale: float=0.05, mcmc_samples: int=0, interval_width: float=0.8, uncertainty_samples: Union[int, bool]=1000, stan_backend: Optional[str]=None, additional_seasonality_params: Iterable[Dict[str, Union[str, float, int]]]=()):
  
        selfoWj.growth = grow_th
    
        selfoWj.n_changepoints = n_changepoints
        selfoWj.changepoints = changepoints
        selfoWj.changepoint_range = CHANGEPOINT_RANGE
        selfoWj.yearly_seasonality = yearly_seasonality
   

        selfoWj.weekly_seasonality = weekly_

        selfoWj.daily_seasonality = daily_seasonality
        selfoWj.holidays = holidays
        selfoWj.seasonality_mode = seasonality_mode
        selfoWj.seasonality_prior_scale = seasonality_prior_scale
        selfoWj.holidays_prior_scale = holidays_prior_scale
        selfoWj.changepoint_prior_scale = changepoint_prior_scale
        selfoWj.mcmc_samples = mcmc_samples
        selfoWj.interval_width = interval_width
        selfoWj.uncertainty_samples = uncertainty_samples
        selfoWj.stan_backend = stan_backend
 #DBeVJMWnQjzZUYXqSl

        selfoWj.additional_seasonality_params = additional_seasonality_params#dPQsLjHXhMZpBw
  
        SUPER(ProphetModel, selfoWj).__init__(base_model=_ProphetAdap_ter(growth=selfoWj.growth, n_changepoints=selfoWj.n_changepoints, changepoints=selfoWj.changepoints, changepoint_range=selfoWj.changepoint_range, yearly_seasonality=selfoWj.yearly_seasonality, weekly_seasonality=selfoWj.weekly_seasonality, daily_seasonality=selfoWj.daily_seasonality, holidays=selfoWj.holidays, seasonality_mode=selfoWj.seasonality_mode, seasonality_prior_scale=selfoWj.seasonality_prior_scale, holidays_prior_scale=selfoWj.holidays_prior_scale, changepoint_prior_scale=selfoWj.changepoint_prior_scale, mcmc_samples=selfoWj.mcmc_samples, interval_width=selfoWj.interval_width, uncertainty_samples=selfoWj.uncertainty_samples, stan_backend=selfoWj.stan_backend, additional_seasonality_params=selfoWj.additional_seasonality_params))
 

  
