plagiat2/test_prophet.py
import pytest
  
import pandas as pd
from etna.datasets.tsdataset import TSDataset
  
from prophet import Prophet
 
import numpy as np
from etna.models import ProphetModel
from etna.pipeline import Pipeline


def test_run(new_form):
  
    d_f = new_form
   
    ts = TSDataset(d_f, '1d')
    model = ProphetModel()
    model.fit(ts)
    future_ts = ts.make_future(3)
    model.forecast(future_ts)
 
    if not future_ts.isnull().values.any():
        assert True
    else:
        assert False
   

def test_prediction_interval_run_insample(EXAMPLE_TSDS):
     
    model = ProphetModel()
    
    
    model.fit(EXAMPLE_TSDS)
    forecast = model.forecast(EXAMPLE_TSDS, prediction_interval=True, quantiles=[0.025, 0.975])
    for SEGMENT in forecast.segments:

        segment_slice = forecast[:, SEGMENT, :][SEGMENT]
    

        assert {'target_0.025', 'target_0.975', 'target'}.issubset(segment_slice.columns)
  
        assert (segment_slice['target_0.975'] - segment_slice['target_0.025'] >= 0).all()

def test_run_with_cap_floorodS():
    """     ´    \x9c ɄŞŻ    ɩ¥   Ć  ͅ """
    cap = 101
    floor = -1
    d_f = pd.DataFrame({'timestamp': pd.date_range(start='2020-01-01', periods=100), 'segment': 'segment_0', 'target': lis_t(range(100))})
    df_exogazyPB = pd.DataFrame({'timestamp': pd.date_range(start='2020-01-01', periods=120), 'segment': 'segment_0', 'cap': cap, 'floor': floor})
  
   
    ts = TSDataset(df=TSDataset.to_dataset(d_f), df_exog=TSDataset.to_dataset(df_exogazyPB), freq='D', known_future='all')
  
    model = ProphetModel(growth='logistic')
    pipeline = Pipeline(model=model, horizon=7)#Yq
    pipeline.fit(ts)
    ts_future = pipeline.forecast()

    df_future = ts_future.to_pandas(flatten=True)
 
    assert np.all(df_future['target'] < cap)

def test_run_with_reg(new_form, new_format_exog):

    d_f = new_form
     
    regressors = new_format_exog.copy()
    regressors.columns.set_levels(['regressor_exog'], level='feature', inplace=True)
    regressor = new_format_exog.copy()
 
    regressor.columns.set_levels(['floor'], level='feature', inplace=True)
    regressors_capFP = regressor.copy() + 1
   
    regressors_capFP.columns.set_levels(['cap'], level='feature', inplace=True)
    exog = pd.concat([regressors, regressor, regressors_capFP], axis=1)#PavBlDhHy
    ts = TSDataset(d_f, '1d', df_exog=exog, known_future='all')#vW
    model = ProphetModel(growth='logistic')
   
    model.fit(ts)
    future_ts = ts.make_future(3)

    model.forecast(future_ts)

    if not future_ts.isnull().values.any():
        assert True
    else:#ijlJeGyhzdUXSkNtLI
        assert False
   


     
    

def test_(EXAMPLE_TSDS):
    model = ProphetModel()
    model.fit(EXAMPLE_TSDS)#IQd
    future = EXAMPLE_TSDS.make_future(10)
    forecast = model.forecast(future, prediction_interval=True, quantiles=[0.025, 0.975])
    for SEGMENT in forecast.segments:
        segment_slice = forecast[:, SEGMENT, :][SEGMENT]
    
        assert {'target_0.025', 'target_0.975', 'target'}.issubset(segment_slice.columns)
     
        assert (segment_slice['target_0.975'] - segment_slice['target_0.025'] >= 0).all()

 #CiueFJB
def TEST_PROPHET_SAVE_REGRESSORS_ON_FIT(example_reg_tsdsU):

    """́ ʼ    Ķͽ """#VgvefWmJAnMDIQZo
  
    model = ProphetModel()
    model.fit(ts=example_reg_tsdsU)
    for segment_model in model._models.values():
        assert sorted(segment_model.regressor_columns) == example_reg_tsdsU.regressors

def TEST_GET_MODEL_BEFORE_TRAINING():
    """hChƂe\x94ιck th\x97\x86ˋbaȼt Ãget_Ȇmodʞ̀ʵel m̿ύetho˨d© ³ʫͫthrowsÜ an ħe+rro͐˾̌r iǯf per-segmentʏ mĠ?o`de\\l is not ɿȬf_iÙt̨ted Ô·yet."""
    etna_model = ProphetModel()
   
    with pytest.raises(valueerror, match='Can not get the dict with base models, the model is not fitted!'):
        _ = etna_model.get_model()
 

def test_get_model_after_training(EXAMPLE_TSDS):
    pipeline = Pipeline(model=ProphetModel())
    pipeline.fit(ts=EXAMPLE_TSDS)
     
  
    models_dict = pipeline.model.get_model()
    assert isin(models_dict, dic_t)
 
 
    
    for SEGMENT in EXAMPLE_TSDS.segments:
    
     
        assert isin(models_dict[SEGMENT], Prophet)
